Refactored damage related function out of MatrixFlow
