+++ src/module/flows/MatrixFlow.ts
@@ -6,11 +6,7 @@ import { SR5Item } from '../item/SR5Item';
 import { MatrixRules } from '../rules/MatrixRules';
 import { OpposedTest } from '../tests/OpposedTest';
 import { SuccessTest } from '../tests/SuccessTest';
-import { DamageType } from '../types/item/Action';
 import { MatrixAttributesType } from '../types/template/Matrix';
-import { TestCreator } from '@/module/tests/TestCreator';
-import { BiofeedbackResistTest } from '@/module/tests/BiofeedbackResistTest';
-import { ResistTestData } from '@/module/tests/flows/ResistTestDataFlow';
 import { Translation } from '../utils/strings';
 
 /**
@@ -106,63 +102,6 @@ export const MatrixFlow = {
         const messageData = { content, speaker: templateData.speaker };
         await ChatMessage.create(messageData);
 
-    },
-
-    /**
-     * Determine if the actor should take bifoeedback damage from the damage taken
-     * @param actor
-     * @param damage
-     */
-    async determineBiofeedbackDamage(actor: SR5Actor, damage: DamageType) {
-        if (actor.takesBiofeedbackDamageFrom(damage)) {
-            const biofeedbackDamage = this.getBiofeedbackResistDamage(actor, damage);
-            const action = DataDefaults.createData('action_roll', {
-                    ...BiofeedbackResistTest._getDefaultTestAction(),
-                    test: 'BiofeedbackResistTest',
-                });
-            const data = TestCreator._prepareTestDataWithAction(action, actor, {
-                ...TestCreator._minimalTestData(),
-                action,
-                type: 'BiofeedbackResistTest',
-            }) as ResistTestData;
-            data.incomingDamage = biofeedbackDamage;
-            data.modifiedDamage = data.incomingDamage;
-
-            const test = TestCreator.fromTestData(data, {source: actor});
-            await test.execute();
-        }
-    },
-
-
-
-    /**
-     * Get the amount of damage an actor needs to resist as Biofeedback Damage
-     * - biofeedback comes from 2 sources: Matrix Damage and Physical (meatspace) damage
-     * - matrix damage becomes biofeedback when someone in VR takes matrix damage laced with Biofeedback
-     * - physical damage becomes biofeedback when a Drone takes damage and someone is Jumped Into it
-     * @param actor
-     * @param damage
-     */
-    getBiofeedbackResistDamage(actor: SR5Actor, damage: DamageType) {
-        const isMatrix = damage.type.value === 'matrix';
-        // if the biofeedback isn't matrix damage, halve the damage to as per SR5 266
-        const damageValue = isMatrix ? damage.value : Math.ceil(damage.value / 2);
-        // if the biofeedback isn't matrix damage, provide an empty string so it is based on the state of hotsim
-        const biofeedback = isMatrix ? damage.biofeedback : '';
-        return MatrixRules.createBiofeedbackDamage(damageValue, actor.isUsingHotSim, biofeedback);
-    },
-
-    /**
-     * Apply dumpshock to an actor and their persona (device).
-     *
-     * Dumpshock can´t be applied to actors not using VR.
-     *
-     * @param actor The actor that is affected by dumpshock.
-     */
-    getDumpshockDamage(actor: SR5Actor) {
-        if (!actor.isUsingVR) return DataDefaults.createData('damage', { type: { base: 'stun', value: 'stun' } });
-
-        return MatrixRules.dumpshockDamage(actor.isUsingHotSim);
     },
 
     /**
